backend/requirements.txt
flask
flask-cors
opencv-python
pillow
numpy
scikit-image
